Update dependencies and improve OpenAPI comments
Refactor comments in `Program.cs` to clarify OpenAPI and Swagger parameters, changing "operations" to "endpoints." Upgrade target framework in `TinyHelpers.Dapper.Sample.csproj` to `net9.0` and update `Microsoft.Data.SqlClient` to version `6.0.1`. Update `Dapper` and `TinyHelpers` versions in `TinyHelpers.Dapper.csproj` to `2.1.66` and `3.2.20`, respectively. In `TinyHelpers.EntityFrameworkCore.csproj`, update `TinyHelpers` to `3.2.20` and `Microsoft.EntityFrameworkCore.Relational` versions to `8.0.13` and `9.0.2` for the respective target frameworks.